zero-copy emit scatter API
Instead of copying application data (stream / DATAGRAM) into the output
packet buffer, and then encrypting that in-place, we can use the seal
scatter API exposed by BoringCrypto to avoid the additional data copy,
by passing the data buffer directly to the crypto API.

The limitation is that the "scattered" payload buffer must be at the end
of the packet, so no additional frames can be queued after that. In the
STREAM case this is not a problem as only a single STREAM frame is
queued anyway, but in the DATAGRAM case we need to avoid pushing more
than one. In addition, PADDING frames can't be added either.